refactor(console): split concept_store's graph-overview and neighborhood queries
concept_store.py owned bounded read queries for the concept graph
database: shared helpers, several small queries, and two large,
self-contained functions -- graph_overview() (~210 lines, full-vs-
sampled visual atlas with its own degree-ranked BFS tree-sampling)
and neighborhood() (~125 lines, bounded single-entity BFS traversal).

Extracted both, verbatim, into their own modules:
- console/concept_graph_overview.py: graph_overview() plus the 5
  ATLAS_* constants only it used.
- console/concept_neighborhood.py: neighborhood().

Both import shared helpers (_connect, _schema_status, _entity) FROM
concept_store.py, so re-exporting the functions back into
concept_store.py would create a circular import. Updated the real
call sites instead: console/endpoints/concepts.py now imports
graph_overview/neighborhood directly; console/core.py was untouched
since it only calls summary()/build_runs(), which didn't move.

tests/test_concept_store.py updated to match -- notably two tests
monkeypatch the ATLAS_* constants directly; their monkeypatch target
moved from concept_store to concept_graph_overview since patching the
old module's attribute no longer reaches the constant the function
actually reads. Verified load-bearing: reverting just the monkeypatch
target back to concept_store fails loudly with AttributeError (the
constants were fully removed, not left behind unused), proving this
wasn't cosmetic.

concept_store.py: 676 -> 334 lines.

Opus-reviewed: PASS, no findings. Confirmed verbatim/behavior-
preserving, no dangling references anywhere in the repo (including
frontend artifacts), and confirmed the constants-placement direction
(co-located with their sole consumer, not left stranded in the parent
module) matches this repo's refactor-middle-ground.md guidance.

Full suite: 800 passed, 29 skipped, same 6 pre-existing sandbox-only
failures as a clean MARM-main checkout (bundled spaCy concept model
unavailable in this environment). marm-console/tests/: 37 passed.
